cmake: fix creation of xmsg shared library
Cannot create a static library with the xmsg protobuf files and then
link it with the core xmsg shared library. But the xmsgproto library is
useful to pass the runtime include directory with the proto headers to
the core sources (with target_include_directories).

Solve by replacing the xmsgproto static library with a CMake 3.x OBJECT
library, which compiles objects files only. Then this objects can be
added to the core xmsg library.

To pass the xmsg proto headers directory, use an INTERFACE only library
called xmsgproto_headers, and link xmsg to it. No objects are actually
linked, but its properties are used by CMake to handle dependencies and
paths (i.e. the include directory).