util: create a single static IP regex object
It seems that creating a std::regex object is a really costly operation,
and using an automatic variable for the IP regex (that is always the
same) is not optimized by GCC, which seems to be the case for Clang (or
the Clang implementation is better).

Move the IP regex to an static variable, to avoid the cost of creating
the object every time the is_ipaddr function is called.