cmake: use FetchContent module to get googletest
The module adds googletest as a subproject on configuration time.
This makes the tests work only with CMake >= 3.11 (when FetchContent was
added).

Now a parent project may also use FetchContent to get googletest for its
own tests, and it can add xMsg as a subproject and build the xMsg tests
without downloading and building googletest twice.